requirements.txt
Pillow==10.0.0
PyPDF2==3.0.1
mutagen==1.47.0
